+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to the "project-colors" extension will be documented in this
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.1.1] - 2025-08-05
+
+### Fixed
+- Fixed random color only being applied at startup when no existing color is present
+- Fixed debounce on project name to work consistently across all settings (especially color picker)
+- Fixed workspace configuration registration error for projectColors.workspaces
+- Fixed random color not being displayed correctly in project settings color picker on startup
+- Fixed status bar color not being applied on startup when "color status bar" is enabled by default
+
 ## [Unreleased]
 
 - Initial release